Tg invest analyz/UI/requirements.txt
altair==4.2.0
click==8.0.4
pandas==1.5.3
plotly==5.9.0
protobuf==3.19.4
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.27.1
SQLAlchemy==2.0.0b1
streamlit==1.5.1
